Switch to getLanguages Promise on android
